Paper Presentation proposal for week 7 (#188)
* Create README.md

* Add assignment proposal for gas optimization presentation

Added a detailed assignment proposal for a presentation on gas optimization versus security in smart contracts, including title, names, deadline, category, and relevance.

* Add README for SBOM verification protocol proposal

This document outlines a proposal for a blockchain-based protocol to ensure the integrity and version verification of Software Bills of Materials (SBOMs). It details the project's objectives, methodology, and relevance to current security challenges in software supply chains.

* Delete contributions/presentation/week5/makari/README.md

* Create README.md

* Enhance README with detailed relevance of solidity-coverage

Expanded the relevance section to emphasize the importance of solidity-coverage in smart contract development and its role in identifying vulnerabilities before deployment.

* Create README.md

* Fix formatting in README.md

---------

Co-authored-by: Sofia Bobadilla <62437209+sofiabobadilla@users.noreply.github.com>